requirements.txt
paho-mqtt==2.1.0
schedule==1.2.2